Moving fixture to conftest.py to avoid lint errors + linting
